@@ -1,45 +1,48 @@
 {
   "name": "@jihyunlab/secret",
-  "version": "1.0.10",
+  "version": "2.0.0",
   "description": "JihyunLab Secret.",
   "license": "MIT",
   "author": "JihyunLab <info@jihyunlab.com> (https://jihyunlab.com)",
   "homepage": "https://jihyunlab.com/library/npm/secret",
   "repository": "https://github.com/jihyunlab/secret",
   "bugs": "https://github.com/jihyunlab/secret/issues",
-  "main": "dist/src/index.js",
-  "types": "dist/src/index.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "scripts": {
     "build": "tsc --build",
-    "start": "tsc --build && node dist/src/index.js",
+    "start": "tsc --build && node dist/index.js",
     "clean": "tsc --build --clean",
-    "lint": "eslint src/**/*.{ts,tsx,js,jsx} && eslint test/**/*.{ts,tsx,js,jsx}",
+    "lint": "eslint \"**/*.{ts,tsx,mts,cts}\"",
+    "prettier": "prettier --check \"**/*.{ts,tsx,mts,cts}\"",
     "test": "jest --coverage --coverageReporters=cobertura"
   },
   "keywords": [
     "secret",
     "crypto",
+    "cipher",
+    "env",
     "encryption",
-    "decryption",
-    "text",
-    "file",
-    "env"
+    "decryption"
   ],
-  "engines": {
-    "node": ">=16"
-  },
   "dependencies": {
-    "@jihyunlab/crypto": "^1.0.11",
-    "dotenv": "^16.3.1",
-    "ignore": "^5.3.0"
+    "@jihyunlab/crypto": "^2.0.10",
+    "dotenv": "^16.4.5"
   },
   "devDependencies": {
-    "@jihyunlab/eslint-config": "^1.1.3",
-    "@jihyunlab/prettier-config": "^1.1.3",
-    "@types/jest": "^29.5.8",
-    "@types/node": "^20.9.0",
+    "@eslint/js": "^9.6.0",
+    "@jihyunlab/eslint-config": "^2.0.8",
+    "@jihyunlab/prettier-config": "^2.0.8",
+    "@types/eslint__js": "^8.42.3",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.14.10",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
     "jest": "^29.7.0",
-    "ts-jest": "^29.1.1",
-    "typescript": "^5.2.2"
+    "prettier": "^3.3.2",
+    "ts-jest": "^29.1.5",
+    "typescript": "^5.5.3",
+    "typescript-eslint": "^7.15.0"
   }
 }
